Extract run time measurement to module
