+++ src/lib.rs
@@ -1,6 +1,4 @@
 //! Compare the performance of two git commits.
-use command::{Config, Validated};
-use std::time::Instant;
 
 /// CLI definition
 pub mod cli;
@@ -8,27 +6,5 @@ pub mod cli;
 /// Command configuration
 pub mod command;
 
-#[must_use]
-/// Record the run time of a validated command configuration.
-pub fn record_runtime(command: &Config<Validated>) -> f64 {
-    let mut invocation = command.to_command();
-
-    let timer = Instant::now();
-    let result = invocation.status();
-    let measurement = timer.elapsed().as_secs_f64();
-
-    match result {
-        Ok(status) if !status.success() => {
-            if let Some(code) = status.code() {
-                // TODO: Change to proper logging
-                println!("Program exited with code {code}");
-            }
-        }
-        Err(error) => {
-            // TODO: Change to proper logging
-            println!("Failed with error {error}");
-        }
-        _ => {}
-    }
-    measurement
-}
+/// Measurement functions
+pub mod measurement;
