+++ src/main.rs
@@ -4,7 +4,7 @@ use std::fmt::Debug;
 use clap::Parser;
 use git_perfdiff::cli::Args;
 use git_perfdiff::command::Config;
-use git_perfdiff::record_runtime;
+use git_perfdiff::measurement::record_runtime;
 
 /// Print an error to stdout and exit with a failure code.
 fn print_error<E: Debug>(error: E) -> ! {
